requirements.txt
streamlit
pandas
numpy
matplotlib
scikit-learn
nltk
requests
python-dotenv

# bisa seperti ini juga
# local packages
# -e .